Add functionality to create and use Confluent environments by name (#20)
* Add functionality to create and use Confluent environments by name

The code is updated to enable creation of new environments in Confluent by name if they don't exist already. It also changes the handling of environment selection from using environment ID to using environment name. The change is intended to simplify the process and improve user experience by letting the user interact with environment names instead of obscure IDs.

* Update argument name for environment

The argument name for specifying the environment has been changed from "--environment" to "--environment-name" for clarification. The change is applied in both the usage message and argument parser, ensuring consistency throughout the script.